requirements.txt
aws-cdk-lib==2.20.0
constructs>=10.0.0,<11.0.0
urllib3<2
aws-cdk.aws-apigatewayv2-alpha
aws-cdk.aws-apigatewayv2-integrations-alpha
